Draw a load duration curve

Creates a load duration curve visualization from the outputs of calc_ldc and summ_ldc as a ggplot object.

draw_ldc(
  .tbl_calc,
  .tbl_summ,
  y_lab = NULL,
  ldc_legend_name = "Allowable Load at State Water Quality Standard",
  measurement_name = "Measurement Value",
  measurement_shape = 21,
  measurement_color = "dodgerblue",
  measurement_alpha = 1,
  summary_name = "Summarized Measured Load",
  summary_stat_shape = 12,
  summary_stat_color = "red",
  label_nudge_y = 0,
  label_font_family = "Arial",
  label_font_size = 3,
  label_break = TRUE
)

Arguments

.tbl_calc
data frame object

.tbl_summ
data frame object

y_lab
optional string for y-axis label name, will be appended with units automatically. default is NULL.

ldc_legend_name
string, provides the name used for the allowable pollutant load line in the legend. required.

measurement_name
string, provides the name used for measured load values in the legend. required.

measurement_shape
aesthetic value passed to the layer plotting measured load values. defaults to 21.

measurement_color
aesthetic value passed to the layer plotting measured load values. defaults to "dodgerblue".

measurement_alpha
aesthetic value passed to the layer plotting measured load values. defaults to 1.

summary_name
string, provides the name used for summary statistic values in the legend. required.

summary_stat_shape
aesthetic value passed to the layer plotting summary statistic values. defaults to 12.

summary_stat_color
aesthetic value passed to the layer plotting summary statistic values. defaults to "red".

label_nudge_y
numeric value to vertically nudge flow category labels. If a log10 transformed scale is being used, a log value is probably appropriate for example log10(1000).

label_font_family
string specifying font family to use in flow category labels.

label_font_size
numeric value specifying font size to use in flow category labels.

label_break
logical, add line breaks to flow cateogry labels. Labels will break at spaces.

Value

ggplot object